+++ openeo/rest/datacube.py
@@ -15,7 +15,7 @@ import pathlib
 import typing
 import warnings
 from builtins import staticmethod
-from typing import Any, Dict, Iterable, List, Optional, Tuple, Union
+from typing import Any, Dict, Iterable, List, Optional, Tuple, Union, Sequence
 
 import numpy as np
 import shapely.geometry
@@ -31,10 +31,7 @@ from openeo.internal.processes.builder import ProcessBuilderBase
 from openeo.internal.warnings import UserDeprecationWarning, deprecated, legacy_alias
 from openeo.metadata import (
     Band,
-    BandDimension,
     CollectionMetadata,
-    SpatialDimension,
-    TemporalDimension,
 )
 from openeo.rest import BandMathException, OpenEoClientException, OperatorException
 from openeo.rest._datacube import (
@@ -67,8 +64,6 @@ if typing.TYPE_CHECKING:
 log = logging.getLogger(__name__)
 
 
-
-
 class DataCube(_ProcessGraphAbstraction):
     """
     Class representing a openEO (raster) data cube.
@@ -80,9 +75,9 @@ class DataCube(_ProcessGraphAbstraction):
     # TODO: set this based on back-end or user preference?
     _DEFAULT_RASTER_FORMAT = "GTiff"
 
-    def __init__(self, graph: PGNode, connection: Connection, metadata: CollectionMetadata = None):
+    def __init__(self, graph: PGNode, connection: Connection, metadata: Optional[CollectionMetadata] = None):
         super().__init__(pgnode=graph, connection=connection)
-        self.metadata = CollectionMetadata.get_or_create(metadata)
+        self.metadata: Optional[CollectionMetadata] = metadata
 
     def process(
         self,
@@ -119,15 +114,24 @@ class DataCube(_ProcessGraphAbstraction):
         # TODO: deprecate `process_with_node``: little added value over just calling DataCube() directly
         return DataCube(graph=pg, connection=self._connection, metadata=metadata or self.metadata)
 
+    def _do_metadata_normalization(self) -> bool:
+        """Do metadata-based normalization/validation of dimension names, band names, ..."""
+        return isinstance(self.metadata, CollectionMetadata)
+
+    def _assert_valid_dimension_name(self, name: str) -> str:
+        if self._do_metadata_normalization():
+            self.metadata.assert_valid_dimension(name)
+        return name
+
     @classmethod
     @openeo_process
     def load_collection(
         cls,
-        collection_id: str,
+        collection_id: Union[str, Parameter],
         connection: Connection = None,
         spatial_extent: Optional[Dict[str, float]] = None,
         temporal_extent: Optional[List[Union[str, datetime.datetime, datetime.date, PGNode]]] = None,
-        bands: Optional[List[str]] = None,
+        bands: Union[None, List[str], Parameter] = None,
         fetch_metadata=True,
         properties: Optional[Dict[str, Union[str, PGNode, typing.Callable]]] = None,
         max_cloud_cover: Optional[float] = None,
@@ -159,17 +163,17 @@ class DataCube(_ProcessGraphAbstraction):
         }
         if isinstance(collection_id, Parameter):
             fetch_metadata = False
-        metadata = connection.collection_metadata(collection_id) if fetch_metadata else None
+        metadata: Optional[CollectionMetadata] = (
+            connection.collection_metadata(collection_id) if fetch_metadata else None
+        )
         if bands:
             if isinstance(bands, str):
                 bands = [bands]
+            elif isinstance(bands, Parameter):
+                metadata = None
             if metadata:
                 bands = [b if isinstance(b, str) else metadata.band_dimension.band_name(b) for b in bands]
                 metadata = metadata.filter_bands(bands)
-            else:
-                # Ensure minimal metadata with best effort band dimension guess (based on `bands` argument).
-                band_dimension = BandDimension("bands", bands=[Band(name=b) for b in bands])
-                metadata = CollectionMetadata({}, dimensions=[band_dimension])
             arguments['bands'] = bands
         if max_cloud_cover:
             properties = properties or {}
@@ -218,17 +222,7 @@ class DataCube(_ProcessGraphAbstraction):
                 'options': options
             }
         )
-
-        metadata = CollectionMetadata(
-            {},
-            dimensions=[
-                SpatialDimension(name="x", extent=[]),
-                SpatialDimension(name="y", extent=[]),
-                TemporalDimension(name="t", extent=[]),
-                BandDimension(name="bands", bands=[Band(name="unknown")]),
-            ],
-        )
-        return cls(graph=pg, connection=connection, metadata=metadata)
+        return cls(graph=pg, connection=connection)
 
     @classmethod
     def _get_temporal_extent(
@@ -289,12 +283,16 @@ class DataCube(_ProcessGraphAbstraction):
 
     @openeo_process
     def filter_bbox(
-            self,
-            *args,
-            west=None, south=None, east=None, north=None,
-            crs=None,
-            base=None, height=None,
-            bbox=None
+        self,
+        *args,
+        west: Optional[float] = None,
+        south: Optional[float] = None,
+        east: Optional[float] = None,
+        north: Optional[float] = None,
+        crs: Optional[Union[int, str]] = None,
+        base: Optional[float] = None,
+        height: Optional[float] = None,
+        bbox: Optional[Sequence[float]] = None,
     ) -> DataCube:
         """
         Limits the data cube to the specified bounding box.
@@ -306,7 +304,7 @@ class DataCube(_ProcessGraphAbstraction):
                 >>> cube.filter_bbox(west=3, south=51, east=4, north=52, crs=4326)
 
             - With a (west, south, east, north) list or tuple
-              (note that EPSG:4326 is the default CRS, so it's not nececarry to specify it explicitly)::
+              (note that EPSG:4326 is the default CRS, so it's not necessary to specify it explicitly)::
 
                 >>> cube.filter_bbox([3, 51, 4, 52])
                 >>> cube.filter_bbox(bbox=[3, 51, 4, 52])
@@ -331,7 +329,10 @@ class DataCube(_ProcessGraphAbstraction):
 
             - With a CRS other than EPSG 4326::
 
-                >>> cube.filter_bbox(west=652000, east=672000, north=5161000, south=5181000, crs=32632)
+                >>> cube.filter_bbox(
+                ... west=652000, east=672000, north=5161000, south=5181000,
+                ... crs=32632
+                ... )
 
             - Deprecated: positional arguments are also supported,
               but follow a non-standard order for legacy reasons::
@@ -339,6 +340,10 @@ class DataCube(_ProcessGraphAbstraction):
                 >>> west, east, north, south = 3, 4, 52, 51
                 >>> cube.filter_bbox(west, east, north, south)
 
+        :param crs: value describing the coordinate reference system.
+            Typically just an int (interpreted as EPSG code, e.g. ``4326``)
+            or a string (handled as authority string, e.g. ``"EPSG:4326"``).
+            See :py:func:`openeo.util.normalize_crs` for more details about additional normalization that is applied to this argument.
         """
         if args and any(k is not None for k in (west, south, east, north, bbox)):
             raise ValueError("Don't mix positional arguments with keyword arguments.")
@@ -428,13 +433,13 @@ class DataCube(_ProcessGraphAbstraction):
         """
         if isinstance(bands, str):
             bands = [bands]
-        bands = [self.metadata.band_dimension.band_name(b) for b in bands]
+        if self._do_metadata_normalization():
+            bands = [self.metadata.band_dimension.band_name(b) for b in bands]
         cube = self.process(
             process_id="filter_bands",
             arguments={"data": THIS, "bands": bands},
+            metadata=self.metadata.filter_bands(bands) if self.metadata else None,
         )
-        if cube.metadata:
-            cube.metadata = cube.metadata.filter_bands(bands)
         return cube
 
     band_filter = legacy_alias(filter_bands, "band_filter", since="0.1.0")
@@ -446,14 +451,14 @@ class DataCube(_ProcessGraphAbstraction):
         :param band: band name, band common name or band index.
         :return: a DataCube instance
         """
-        band_index = self.metadata.get_band_index(band)
-        return self.reduce_bands(reducer=PGNode(
-            process_id='array_element',
-            arguments={
-                'data': {'from_parameter': 'data'},
-                'index': band_index
-            },
-        ))
+        if self._do_metadata_normalization():
+            band = self.metadata.band_dimension.band_index(band)
+        return self.reduce_bands(
+            reducer=PGNode(
+                process_id="array_element",
+                arguments={"data": {"from_parameter": "data"}, "index": band},
+            )
+        )
 
     @openeo_process
     def resample_spatial(
@@ -829,6 +834,9 @@ class DataCube(_ProcessGraphAbstraction):
     ) -> Union[dict, Parameter, PGNode]:
         """
         Convert input to a geometry as "geojson" subtype object.
+
+        :param crs: value that encodes a coordinate reference system.
+            See :py:func:`openeo.util.normalize_crs` for more details about additional normalization that is applied to this argument.
         """
         if isinstance(geometry, (str, pathlib.Path)):
             # Assumption: `geometry` is path to polygon is a path to vector file at backend.
@@ -877,7 +885,7 @@ class DataCube(_ProcessGraphAbstraction):
         ],
         reducer: Union[str, typing.Callable, PGNode],
         target_dimension: Optional[str] = None,
-        crs: Optional[str] = None,
+        crs: Optional[Union[int, str]] = None,
         context: Optional[dict] = None,
         # TODO arguments: target dimension, context
     ) -> VectorCube:
@@ -903,7 +911,9 @@ class DataCube(_ProcessGraphAbstraction):
 
         :param target_dimension: The new dimension name to be used for storing the results.
         :param crs: The spatial reference system of the provided polygon.
-            By default longitude-latitude (EPSG:4326) is assumed.
+            By default, longitude-latitude (EPSG:4326) is assumed.
+            See :py:func:`openeo.util.normalize_crs` for more details about additional normalization that is applied to this argument.
+
         :param context: Additional data to be passed to the reducer process.
 
             .. note:: this ``crs`` argument is a non-standard/experimental feature, only supported by specific back-ends.
@@ -1070,7 +1080,7 @@ class DataCube(_ProcessGraphAbstraction):
         arguments = {
             "data": THIS,
             "process": process,
-            "dimension": self.metadata.assert_valid_dimension(dimension),
+            "dimension": self._assert_valid_dimension_name(dimension),
         }
         if target_dimension is not None:
             arguments["target_dimension"] = target_dimension
@@ -1115,15 +1125,18 @@ class DataCube(_ProcessGraphAbstraction):
             process=reducer, parent_parameters=["data", "context"], connection=self.connection
         )
 
-        return self.process_with_node(ReduceNode(
-            process_id=process_id,
-            data=self,
-            reducer=reducer,
-            dimension=self.metadata.assert_valid_dimension(dimension),
-            context=context,
-            # TODO #123 is it (still) necessary to make "band" math a special case?
-            band_math_mode=band_math_mode
-        ), metadata=self.metadata.reduce_dimension(dimension_name=dimension))
+        return self.process_with_node(
+            ReduceNode(
+                process_id=process_id,
+                data=self,
+                reducer=reducer,
+                dimension=self._assert_valid_dimension_name(dimension),
+                context=context,
+                # TODO #123 is it (still) necessary to make "band" math a special case?
+                band_math_mode=band_math_mode,
+            ),
+            metadata=self.metadata.reduce_dimension(dimension_name=dimension) if self.metadata else None,
+        )
 
     # @openeo_process
     def chunk_polygon(
@@ -1175,7 +1188,11 @@ class DataCube(_ProcessGraphAbstraction):
 
         :param reducer: "child callback" function, see :ref:`callbackfunctions`
         """
-        return self.reduce_dimension(dimension=self.metadata.band_dimension.name, reducer=reducer, band_math_mode=True)
+        return self.reduce_dimension(
+            dimension=self.metadata.band_dimension.name if self.metadata else "bands",
+            reducer=reducer,
+            band_math_mode=True,
+        )
 
     def reduce_temporal(self, reducer: Union[str, PGNode, typing.Callable, UDF]) -> DataCube:
         """
@@ -1183,7 +1200,10 @@ class DataCube(_ProcessGraphAbstraction):
 
         :param reducer: "child callback" function, see :ref:`callbackfunctions`
         """
-        return self.reduce_dimension(dimension=self.metadata.temporal_dimension.name, reducer=reducer)
+        return self.reduce_dimension(
+            dimension=self.metadata.temporal_dimension.name if self.metadata else "t",
+            reducer=reducer,
+        )
 
     @deprecated(
         "Use :py:meth:`reduce_bands` with :py:class:`UDF <openeo.rest._datacube.UDF>` as reducer.",
@@ -1213,7 +1233,7 @@ class DataCube(_ProcessGraphAbstraction):
         return self.process(
             process_id="add_dimension",
             arguments=dict_no_none({"data": self, "name": name, "label": label, "type": type}),
-            metadata=self.metadata.add_dimension(name=name, label=label, type=type)
+            metadata=self.metadata.add_dimension(name=name, label=label, type=type) if self.metadata else None,
         )
 
     @openeo_process
@@ -1231,7 +1251,7 @@ class DataCube(_ProcessGraphAbstraction):
         return self.process(
             process_id="drop_dimension",
             arguments={"data": self, "name": name},
-            metadata=self.metadata.drop_dimension(name=name),
+            metadata=self.metadata.drop_dimension(name=name) if self.metadata else None,
         )
 
     @deprecated(
@@ -1486,9 +1506,12 @@ class DataCube(_ProcessGraphAbstraction):
 
         :return: a DataCube instance
         """
-        if target_band is None:
+        if self.metadata is None:
+            metadata = None
+        elif target_band is None:
             metadata = self.metadata.reduce_dimension(self.metadata.band_dimension.name)
         else:
+            # TODO: first drop "bands" dim and re-add it with single "ndvi" band
             metadata = self.metadata.append_band(Band(name=target_band, common_name="ndvi"))
         return self.process(
             process_id="ndvi",
@@ -1508,16 +1531,16 @@ class DataCube(_ProcessGraphAbstraction):
 
         :return: A new datacube with the dimension renamed.
         """
-        if target in self.metadata.dimension_names():
+        if self._do_metadata_normalization() and target in self.metadata.dimension_names():
             raise ValueError('Target dimension name conflicts with existing dimension: %s.' % target)
         return self.process(
             process_id="rename_dimension",
             arguments=dict_no_none(
                 data=THIS,
-                source=self.metadata.assert_valid_dimension(source),
+                source=self._assert_valid_dimension_name(source),
                 target=target,
             ),
-            metadata=self.metadata.rename_dimension(source, target),
+            metadata=self.metadata.rename_dimension(source, target) if self.metadata else None,
         )
 
     @openeo_process
@@ -1535,11 +1558,11 @@ class DataCube(_ProcessGraphAbstraction):
             process_id="rename_labels",
             arguments=dict_no_none(
                 data=THIS,
-                dimension=self.metadata.assert_valid_dimension(dimension),
+                dimension=self._assert_valid_dimension_name(dimension),
                 target=target,
                 source=source,
             ),
-            metadata=self.metadata.rename_labels(dimension, target, source),
+            metadata=self.metadata.rename_labels(dimension, target, source) if self.metadata else None,
         )
 
     @openeo_process(mode="apply")
@@ -1656,12 +1679,20 @@ class DataCube(_ProcessGraphAbstraction):
         arguments = {"cube1": self, "cube2": other}
         if overlap_resolver:
             arguments["overlap_resolver"] = build_child_callback(overlap_resolver, parent_parameters=["x", "y"])
-        # Minimal client side metadata merging
-        merged_metadata = self.metadata
-        if self.metadata.has_band_dimension() and isinstance(other, DataCube) and other.metadata.has_band_dimension():
+        if (
+            self.metadata
+            and self.metadata.has_band_dimension()
+            and isinstance(other, DataCube)
+            and other.metadata
+            and other.metadata.has_band_dimension()
+        ):
+            # Minimal client side metadata merging
+            merged_metadata = self.metadata
             for b in other.metadata.band_dimension.bands:
                 if b not in merged_metadata.bands:
                     merged_metadata = merged_metadata.append_band(b)
+        else:
+            merged_metadata = None
         # Overlapping bands without overlap resolver will give an error in the backend
         if context:
             arguments["context"] = context
@@ -1741,8 +1772,7 @@ class DataCube(_ProcessGraphAbstraction):
         :return: a :py:class:`~openeo.rest.vectorcube.VectorCube`
         """
         pg_node = PGNode(process_id="raster_to_vector", arguments={"data": self})
-        # TODO: properly update metadata (e.g. "geometry" dimension) related to #457
-        return VectorCube(pg_node, connection=self._connection, metadata=self.metadata)
+        return VectorCube(pg_node, connection=self._connection)
 
     ####VIEW methods #######
 
@@ -2316,9 +2346,10 @@ class DataCube(_ProcessGraphAbstraction):
 
         :param dimension: The name of the dimension to get the labels for.
         """
-        dimension_names = self.metadata.dimension_names()
-        if dimension_names and dimension not in dimension_names:
-            raise ValueError(f"Invalid dimension name {dimension!r}, should be one of {dimension_names}")
+        if self._do_metadata_normalization():
+            dimension_names = self.metadata.dimension_names()
+            if dimension_names and dimension not in dimension_names:
+                raise ValueError(f"Invalid dimension name {dimension!r}, should be one of {dimension_names}")
         return self.process(process_id="dimension_labels", arguments={"data": THIS, "dimension": dimension})
 
     @openeo_process