+++ CHANGELOG.md
@@ -10,9 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Allow shorthand format for temporal extent to specify entire year or month ([#421](https://github.com/Open-EO/openeo-python-client/issues/421))
+- Support parameterized `bands` in `load_collection` ([#471](https://github.com/Open-EO/openeo-python-client/issues/471))
+- Support shorthand format for temporal extents to specify entire year or month ([#421](https://github.com/Open-EO/openeo-python-client/issues/421))
 
 ### Changed
+
+- Before doing user-defined process (UDP) listing/creation: verify that back-end supports that (through openEO capabilities document) to improve error message.
+- Skip metadata-based normalization/validation and stop showing unhelpful warnings/errors
+  like "No cube:dimensions metadata" or "Invalid dimension"
+  when no metadata is available client-side anyway (e.g. when using `datacube_from_process`, parameterized cube building, ...).
+  ([#442](https://github.com/Open-EO/openeo-python-client/issues/442))
 
 ### Removed